# api_specification.yaml
openapi: 3.0.3
info:
  title: API do Exemplo da Aula de Programação III
  description: Definição exemplificativa de uma API em OpenAPI 3 para o exercício prático das aulas da UC de Programação III do curso de Multimédia do ISMT.
  termsOfService: http://swagger.io/terms/
  contact:
    email: ismt@ismt.pt
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
servers:
  - url: http://localhost:5000/api/v1
    description: Servidor local
security:
  - bearerAuth: []
...
paths:
...
  /teachers:
    get:
      tags:
        - teachers
      summary: Retorna todos os professores
      operationId: getAllTeachers
      responses:
        "200":
          description: Lista de professores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teacher"
        "401":
          description: Não autorizado
        "500":
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                message: Erro interno no servidor
                code: 500
  /teacher:
    post:
      tags:
        - teachers
      summary: Cria um novo professor
      operationId: createTeacher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Teacher"
      responses:
        "201":
          description: Professor criado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                message: Erro interno no servidor
                code: 500
  /teacher/{id}:
    get:
      tags:
        - teachers
      summary: Retorna um professor pelo ID
      operationId: getTeacherById
      parameters:
        - in: path
          name: id
          required: true
          description: ID do professor a ser retornado
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Professor encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
        "404":
          description: Professor não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                message: Erro interno no servidor
                code: 500
    put:
      tags:
        - teachers
      summary: Atualiza um professor pelo ID
      operationId: updateTeacher
      parameters:
        - in: path
          name: id
          required: true
          description: ID do professor a ser atualizado
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Teacher"
      responses:
        "200":
          description: Professor atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
        "404":
          description: Professor não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                message: Erro interno no servidor
                code: 500
    delete:
      tags:
        - teachers
      summary: Remove um professor pelo ID
      operationId: deleteTeacher
      parameters:
        - in: path
          name: id
          required: true
          description: ID do professor a ser removido
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "204":
          description: Professor removido com sucesso
        "404":
          description: Professor não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                message: Erro interno no servidor
                code: 500
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Teacher:
      required:
        - id
        - name
        - nif
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Chave primária do professor
          example: 1
        name:
          type: string
          description: Nome completo do professor
          example: Maria Sousa
        email:
          type: string
          description: Endereço de email do professor
          example: maria@ismt.pt
        nif:
          type: integer
          format: int64
          description: NIF do professor (9 dígitos)
          example: 987654321
          minimum: 000000001
          maximum: 999999999
        subjects:
          type: array
          description: Todas as disciplinas que o professor leciona
          items:
            $ref: '#/components/schemas/Subject'
    Subject:
      required:
        - id
        - name
        - totalHours
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Chave primária da disciplina
          example: 10
        name:
          type: string
          description: Nome da disciplina
          example: Programação III
        totalHours:
          type: number
          format: double
          description: Total de horas letivas da disciplina
          example: 64
        teachers:
          type: array
          description: Todos os professores que lecionam a disciplina
          items:
            $ref: '#/components/schemas/Teacher'
...
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: integer
      example:
        message: Recurso não encontrado
        code: 404